[task scripts] Files that match/simplify the current npm scripts
Demo commit only - functionality not implemented

This approach creates scripting files in `bin` that call the relevant tasks. They correspond to our current npm scripts.

For example, `bin/dev.js` contains the tasks for `npm start` and `bin/deploy.js` contains the tasks for `npm run heroku`.

Any functionality that's only used within one script can be defined there - we only watch files when we run `npm start`, for instance.

There's a chance that naive functions could noticeably slow down the build (for example, if we compile the JS, save it to disk, then copy it to another location). If this proves to be the case, we could look into streaming or other solutions to avoid too much I/O.

VARIATION: Scripting takes place in `.sh` files. We already have a few of these in `/bin`, but the assumption is that most devs working with these files will be more comfortable working with JavaScript.