[npm scripts] Add a bunch of npm scripts
Non-functional demo

Creates a bunch of `npm` scripts which are then used to perform the necessary tasks.

In this example, I've assumed similar scripts would share a script file, and others would have their own.

Despite my earlier reservations about this, it does feel neat in that we can compose tasks via npm. A flaw is that an individual task may not do anything noticeably useful without building on other tasks (copying compiled assets, for example). There's also a big ol' load of text to parse. Ordering the scripts so that the main relevant ones are listed first, then the subscripts alphabetically may help.

In the case of scripts which involve multiple subscripts, we may want to investigate something like `npm-run-all` or the like.